+++ Sources/TraceLog/Formatters/ByteOutputFormatter.swift
@@ -1,5 +1,5 @@
 ///
-///  TextOutputFormatter.swift
+///  ByteOutputFormatter.swift
 ///
 ///  Copyright 2018 Tony Stone
 ///
@@ -21,6 +21,6 @@ import Foundation
 
 /// A formatter type for formating the output of a `Writer` type.
 ///
-public protocol TextOutputFormatter {
-    func text(from timestamp: Double, level: LogLevel, tag: String, message: String, runtimeContext: RuntimeContext, staticContext: StaticContext) -> TextOutputStreamable?
+public protocol ByteOutputFormatter {
+    func bytes(from timestamp: Double, level: LogLevel, tag: String, message: String, runtimeContext: RuntimeContext, staticContext: StaticContext) -> [UInt8]?
 }